+++ website/next-sitemap.config.js
@@ -1,7 +1,7 @@
 /** @type {import('next-sitemap').IConfig} */
-
-module.exports = {
+export default {
   siteUrl: process.env.SITE_URL || 'https://the-guild.dev/graphql/inspector',
   generateIndexSitemap: false,
   exclude: ['*/_meta'],
+  output: 'export',
 };