+++ packages/commands/commands/CHANGELOG.md
@@ -1,5 +1,86 @@
 # @graphql-inspector/commands
 
+## 5.0.4
+
+### Patch Changes
+
+- [#2740](https://github.com/kamilkisiela/graphql-inspector/pull/2740)
+  [`12a2208`](https://github.com/kamilkisiela/graphql-inspector/commit/12a2208748d3ef987400e298eb796e9139357a53)
+  Thanks [@TuvalSimha](https://github.com/TuvalSimha)! - dependencies updates:
+  - Updated dependency
+    [`@graphql-inspector/loaders@^4.0.0` ↗︎](https://www.npmjs.com/package/@graphql-inspector/loaders/v/4.0.0)
+    (from `^4.0.4`, in `peerDependencies`)
+
+## 5.0.3
+
+### Patch Changes
+
+- [#2738](https://github.com/kamilkisiela/graphql-inspector/pull/2738)
+  [`1600e72`](https://github.com/kamilkisiela/graphql-inspector/commit/1600e723a3bfb9941674d31b49da90e98463b20f)
+  Thanks [@TuvalSimha](https://github.com/TuvalSimha)! - dependencies updates:
+  - Updated dependency
+    [`@graphql-inspector/config@^4.0.0` ↗︎](https://www.npmjs.com/package/@graphql-inspector/config/v/4.0.0)
+    (from `^4.0.2`, in `peerDependencies`)
+  - Updated dependency
+    [`@graphql-inspector/loaders@^4.0.0` ↗︎](https://www.npmjs.com/package/@graphql-inspector/loaders/v/4.0.0)
+    (from `^4.0.4`, in `peerDependencies`)
+- Updated dependencies
+  [[`1600e72`](https://github.com/kamilkisiela/graphql-inspector/commit/1600e723a3bfb9941674d31b49da90e98463b20f)]:
+  - @graphql-inspector/loaders@4.0.4
+
+## 5.0.2
+
+### Patch Changes
+
+- [#2736](https://github.com/kamilkisiela/graphql-inspector/pull/2736)
+  [`87b25d3`](https://github.com/kamilkisiela/graphql-inspector/commit/87b25d32b298efb91213f593ea3fd98f9f4992c4)
+  Thanks [@TuvalSimha](https://github.com/TuvalSimha)! - dependencies updates:
+  - Updated dependency
+    [`@graphql-inspector/config@^4.0.2` ↗︎](https://www.npmjs.com/package/@graphql-inspector/config/v/4.0.2)
+    (from `^5.0.1`, in `peerDependencies`)
+  - Updated dependency
+    [`@graphql-inspector/loaders@^4.0.3` ↗︎](https://www.npmjs.com/package/@graphql-inspector/loaders/v/4.0.3)
+    (from `^5.0.1`, in `peerDependencies`)
+- Updated dependencies
+  [[`87b25d3`](https://github.com/kamilkisiela/graphql-inspector/commit/87b25d32b298efb91213f593ea3fd98f9f4992c4)]:
+  - @graphql-inspector/loaders@4.0.4
+
+## 5.0.1
+
+### Patch Changes
+
+- [#2733](https://github.com/kamilkisiela/graphql-inspector/pull/2733)
+  [`087d3ec`](https://github.com/kamilkisiela/graphql-inspector/commit/087d3ec2056812ee3682e8398d3d255bd7255f5f)
+  Thanks [@TuvalSimha](https://github.com/TuvalSimha)! - dependencies updates:
+  - Updated dependency
+    [`@graphql-inspector/config@^4.0.0` ↗︎](https://www.npmjs.com/package/@graphql-inspector/config/v/4.0.0)
+    (from `5.0.0`, in `peerDependencies`)
+  - Updated dependency
+    [`@graphql-inspector/loaders@^4.0.0` ↗︎](https://www.npmjs.com/package/@graphql-inspector/loaders/v/4.0.0)
+    (from `5.0.0`, in `peerDependencies`)
+- Updated dependencies
+  [[`087d3ec`](https://github.com/kamilkisiela/graphql-inspector/commit/087d3ec2056812ee3682e8398d3d255bd7255f5f),
+  [`087d3ec`](https://github.com/kamilkisiela/graphql-inspector/commit/087d3ec2056812ee3682e8398d3d255bd7255f5f)]:
+  - @graphql-inspector/loaders@5.0.1
+  - @graphql-inspector/config@5.0.1
+
+## 5.0.0
+
+### Major Changes
+
+- [#2719](https://github.com/kamilkisiela/graphql-inspector/pull/2719)
+  [`50874f2`](https://github.com/kamilkisiela/graphql-inspector/commit/50874f2d35d0d5ffda133bb35bb7ce0f8cdb64c3)
+  Thanks [@TuvalSimha](https://github.com/TuvalSimha)! - Upgrade the Node version from 16 to 18, as
+  the minimum required version is now 18. Additionally, ensure that GraphQL Yoga is upgraded to
+  version 5.
+
+### Patch Changes
+
+- Updated dependencies
+  [[`50874f2`](https://github.com/kamilkisiela/graphql-inspector/commit/50874f2d35d0d5ffda133bb35bb7ce0f8cdb64c3)]:
+  - @graphql-inspector/loaders@5.0.0
+  - @graphql-inspector/config@5.0.0
+
 ## 4.0.3
 
 ### Patch Changes