+++ packages/commands/serve/CHANGELOG.md
@@ -1,5 +1,85 @@
 # @graphql-inspector/serve-command
 
+## 5.0.6
+
+### Patch Changes
+
+- [#2812](https://github.com/kamilkisiela/graphql-inspector/pull/2812)
+  [`eba0184`](https://github.com/kamilkisiela/graphql-inspector/commit/eba01849686f6e2f86e4767c36ed397638d869c5)
+  Thanks [@renovate](https://github.com/apps/renovate)! - dependencies updates:
+  - Updated dependency
+    [`graphql-yoga@5.7.0` ↗︎](https://www.npmjs.com/package/graphql-yoga/v/5.7.0) (from `5.6.3`, in
+    `dependencies`)
+
+## 5.0.5
+
+### Patch Changes
+
+- [#2747](https://github.com/kamilkisiela/graphql-inspector/pull/2747)
+  [`9c36138`](https://github.com/kamilkisiela/graphql-inspector/commit/9c3613806803343af8e3081ffc6fea243cea72b2)
+  Thanks [@TuvalSimha](https://github.com/TuvalSimha)! - dependencies updates:
+  - Updated dependency
+    [`graphql-yoga@5.3.1` ↗︎](https://www.npmjs.com/package/graphql-yoga/v/5.3.1) (from `5.3.0`, in
+    `dependencies`)
+- Updated dependencies
+  [[`9c36138`](https://github.com/kamilkisiela/graphql-inspector/commit/9c3613806803343af8e3081ffc6fea243cea72b2)]:
+  - @graphql-inspector/logger@5.0.1
+
+## 5.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  [[`12a2208`](https://github.com/kamilkisiela/graphql-inspector/commit/12a2208748d3ef987400e298eb796e9139357a53)]:
+  - @graphql-inspector/commands@5.0.4
+
+## 5.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  [[`1600e72`](https://github.com/kamilkisiela/graphql-inspector/commit/1600e723a3bfb9941674d31b49da90e98463b20f)]:
+  - @graphql-inspector/commands@5.0.3
+
+## 5.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  [[`87b25d3`](https://github.com/kamilkisiela/graphql-inspector/commit/87b25d32b298efb91213f593ea3fd98f9f4992c4)]:
+  - @graphql-inspector/commands@5.0.2
+
+## 5.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  [[`087d3ec`](https://github.com/kamilkisiela/graphql-inspector/commit/087d3ec2056812ee3682e8398d3d255bd7255f5f)]:
+  - @graphql-inspector/commands@5.0.1
+
+## 5.0.0
+
+### Major Changes
+
+- [#2719](https://github.com/kamilkisiela/graphql-inspector/pull/2719)
+  [`50874f2`](https://github.com/kamilkisiela/graphql-inspector/commit/50874f2d35d0d5ffda133bb35bb7ce0f8cdb64c3)
+  Thanks [@TuvalSimha](https://github.com/TuvalSimha)! - Upgrade the Node version from 16 to 18, as
+  the minimum required version is now 18. Additionally, ensure that GraphQL Yoga is upgraded to
+  version 5.
+
+### Patch Changes
+
+- [#2719](https://github.com/kamilkisiela/graphql-inspector/pull/2719)
+  [`50874f2`](https://github.com/kamilkisiela/graphql-inspector/commit/50874f2d35d0d5ffda133bb35bb7ce0f8cdb64c3)
+  Thanks [@TuvalSimha](https://github.com/TuvalSimha)! - dependencies updates:
+  - Updated dependency
+    [`graphql-yoga@5.3.0` ↗︎](https://www.npmjs.com/package/graphql-yoga/v/5.3.0) (from `4.0.3`, in
+    `dependencies`)
+- Updated dependencies
+  [[`50874f2`](https://github.com/kamilkisiela/graphql-inspector/commit/50874f2d35d0d5ffda133bb35bb7ce0f8cdb64c3)]:
+  - @graphql-inspector/commands@5.0.0
+  - @graphql-inspector/logger@5.0.0
+
 ## 4.0.3
 
 ### Patch Changes